requirements.txt
openai>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
websockets>=12.0
elevenlabs>=0.2.26
pygame>=2.5.0
pyaudio>=0.2.11
wave>=0.0.2
python-multipart>=0.0.6
aiofiles>=23.0.0
firebase-admin>=7.0.0